Actualizacion en todo el Admin

+++ src/app/modules/test-question/listar/listar.component.ts
@@ -1,6 +1,7 @@
 import { InterfaceQuestion } from '@/app/core/interfaces/interface-question';
 import { FilterTablesPipe } from '@/app/shared/pipes/filter-tables.pipe';
 import { TestQuestionService } from '@/app/shared/services/api/test-question.service';
+import { NotificationsService } from '@/app/shared/services/utils/notifications.service';
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 
@@ -8,39 +9,56 @@ import { Router } from '@angular/router';
   selector: 'app-listar',
   templateUrl: './listar.component.html',
   styleUrls: ['./listar.component.css'],
-  providers: [FilterTablesPipe]
+  providers: [FilterTablesPipe],
 })
 export class ListarComponent {
-
   question: InterfaceQuestion[] = [];
   public search: string = '';
 
   constructor(
     private questionService: TestQuestionService,
-    public filterTable: FilterTablesPipe,
+    private notification: NotificationsService,
     private router: Router
-  ){}
+  ) {}
 
-  ngOnInit():void{
-    this.questionService.getAllQuestion().subscribe(
-      res => {
-        const { message, data } = res;
-        this.question = data;
-        console.log(message)
-      }
-    )
+  ngOnInit(): void {
+    this.questionService.getAllQuestion().subscribe((res) => {
+      const { message, data } = res;
+      this.question = data;
+      console.log(message);
+    });
   }
 
-  delete(id:any){
-    this.questionService.deleteQuestion(id).subscribe(
-      res => {
-        const { message } = res;
-        console.log(message)
-        this.ngOnInit();
-      }, err => {
-        console.log(err)
-      }
-    )
+  delete(id: any) {
+    this.notification
+      .showConfirm(
+        'warning',
+        'Peligro',
+        'Estas seguro de eliminar la pregunta?',
+        'Si, eliminar!',
+        'No, cancelar!'
+      )
+      .then((result) => {
+        if (result.isConfirmed) {
+          this.questionService.deleteQuestion(id).subscribe(
+            (res) => {
+              const { message } = res;
+              this.notification.showSuccess(
+                'Éxito',
+                'Pregunta eliminada correctamente'
+              );
+              console.log(message);
+              this.ngOnInit();
+            },
+            (err) => {
+              console.log(err);
+              this.notification.showError(
+                'Error',
+                'No se pudo eliminar la pregunta'
+              );
+            }
+          );
+        }
+      });
   }
-
 }
